+++ cynic-codegen/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod enum_derive;
 pub mod fragment_arguments_derive;
 pub mod fragment_derive;
 pub mod inline_fragments_derive;